Change refactor namespace to tests
