+++ PVTPackage/source/tests/deserializers/BlackOilDeadOilApiInputs.cpp
@@ -18,7 +18,7 @@
 
 namespace PVTPackage
 {
-namespace refactor
+namespace tests
 {
 namespace pds
 {
