+++ PVTPackage/source/tests/deserializers/BlackOilDeadOilApiInputs.hpp
@@ -21,7 +21,7 @@
 
 namespace PVTPackage
 {
-namespace refactor
+namespace tests
 {
 namespace pds
 {
