+++ PVTPackage/source/tests/deserializers/CompositionalApiInputs.cpp
@@ -19,7 +19,7 @@
 
 namespace PVTPackage
 {
-namespace refactor
+namespace tests
 {
 namespace pds
 {
