+++ PVTPackage/source/tests/deserializers/CompositionalApiInputs.hpp
@@ -21,7 +21,7 @@
 
 namespace PVTPackage
 {
-namespace refactor
+namespace tests
 {
 namespace pds
 {
