+++ PVTPackage/source/tests/deserializers/MultiphaseSystemProperties.hpp
@@ -26,7 +26,7 @@
 
 namespace PVTPackage
 {
-namespace refactor
+namespace tests
 {
 namespace pds
 {
