+++ PVTPackage/source/tests/deserializers/PropertyAndDerivatives.hpp
@@ -20,7 +20,7 @@
 #include <nlohmann/json.hpp>
 
 namespace PVTPackage {
-namespace refactor {
+namespace tests {
 namespace pds {
 
 void from_json( const nlohmann::json & j,
@@ -30,7 +30,7 @@ void from_json( const nlohmann::json & j,
                 pds::VectorPropertyAndDerivatives< double > & v );
 
 } // end of namespace pds
-} // end of namespace refactor
+} // end of namespace tests
 } // end of namespace PVTPackage
 
 #endif // PVTPACKAGE_DESERIALIZERS_PROPERTYANDDERIVATIVES_HPP
