+++ PVTPackage/source/tests/passiveDataStructures/BlackOilDeadOilApiInputs.cpp
@@ -16,7 +16,7 @@
 
 namespace PVTPackage
 {
-namespace refactor
+namespace tests
 {
 namespace pds
 {
