+++ PVTPackage/source/tests/passiveDataStructures/MultiphaseSystemProperties.cpp
@@ -16,7 +16,7 @@
 
 namespace PVTPackage
 {
-namespace refactor
+namespace tests
 {
 namespace pds
 {
