+++ PVTPackage/source/tests/passiveDataStructures/MultiphaseSystemProperties.hpp
@@ -23,7 +23,7 @@
 
 namespace PVTPackage
 {
-namespace refactor
+namespace tests
 {
 namespace pds
 {
