+++ PVTPackage/source/tests/testParallel.cpp
@@ -30,7 +30,7 @@
 
 namespace PVTPackage
 {
-namespace refactor
+namespace tests
 {
 
 using json = nlohmann::json;
