+++ PVTPackage/source/tests/testPublicApi.cpp
@@ -41,7 +41,7 @@
 
 namespace PVTPackage
 {
-namespace refactor
+namespace tests
 {
 
 using json = nlohmann::json;
